refactor: 스프링부트 애플리케이션명 변경
