+++ src/main/java/be/dash/dashserver/DashServerApplication.java
@@ -1,13 +1,13 @@
-package be.doomchit.server;
+package be.dash.dashserver;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ServerApplication {
+public class DashServerApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ServerApplication.class, args);
+		SpringApplication.run(DashServerApplication.class, args);
 	}
 
 }
