+++ src/test/java/be/dash/dashserver/ServerApplicationTests.java
@@ -1,4 +1,4 @@
-package be.doomchit.server;
+package be.dash.dashserver;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
